По замечаниям 1: Поправки под requirejs, исправление того что сломал рибейз
