+++ src/picture.js
@@ -6,31 +6,20 @@
 define(function() {
   var IMAGE_LOAD_TIMEOUT = 10000;
 
-  var elemTemplate = document.querySelector('#picture-template');
-  var elemToClone;
-
-  if ('content' in elemTemplate) {
-    elemToClone = elemTemplate.content.querySelector('.picture');
-  } else {
-    elemToClone = elemTemplate.querySelector('.picture');
-  }
-
-  var getPictureElement = function(data, container) {
-    var elem = elemToClone.cloneNode(true);
+  var getPictureElement = function(data, elem) {
     var imgElem = elem.querySelector('img');
     var tileImage = new Image();
     var tileTimeout;
 
-    container.appendChild(elem);
-
     tileImage.onload = function() {
       clearTimeout(tileTimeout);
-      imgElem.src = tileImage.src;
-      imgElem.width = 182;
-      imgElem.height = 182;
+      elem.replaceChild(tileImage, imgElem);
+      tileImage.width = 182;
+      tileImage.height = 182;
     };
 
     tileImage.onerror = function() {
+      clearTimeout(tileTimeout);
       elem.classList.add('picture-load-failure');
     };
